w4/requirements.txt
Flask==3.0.3
flask-swagger-ui==5.21.0
PyYAML==6.0.2
